% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.webmd.primcare-package.R
\docType{package}
\name{dc.webmd.primcare-package}
\alias{dc.webmd.primcare}
\alias{dc.webmd.primcare-package}
\title{dc.webmd.primcare: Social Data Commons: Health, Behavior, Diet: Primary Care Access Scores}
\description{
Allows user to easily get YOUR_DATA data from the SDAD Data Commons. Data can be provided directly or in a file.
}
\author{
\strong{Maintainer}: Cesar Montalvo \email{cpm9w@virginia.edu}

Authors:
\itemize{
  \item Social and Decision Analytics Division (SDAD), Biocomplexity Institute, University of Virginia
}

}
\keyword{internal}
